Move classes into separate packages. Fix order activity.
