+++ app/src/main/AndroidManifest.xml
@@ -21,25 +21,25 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".BasketActivity"
+            android:name=".ui.navigation.BasketActivity"
             android:label="@string/title_activity_basket"
             android:theme="@style/AppTheme" >
 
 
         </activity>
         <activity
-            android:name=".AboutDelevery"
+            android:name=".ui.navigation.AboutDelevery"
             android:label="@string/title_activity_about_delevery"
             android:theme="@style/AppTheme" />
-        <activity android:name=".InfoItemsActivity" />
-        <activity android:name=".ProfileActivity" />
-        <activity android:name=".SalesActivity" />
+        <activity android:name=".ui.navigation.InfoItemsActivity" />
+        <activity android:name=".ui.navigation.ProfileActivity" />
+        <activity android:name=".ui.navigation.SalesActivity" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".ui.navigation.LoginActivity"
             android:label="@string/title_activity_login" />
-        <activity android:name=".SingUpActivity" />
-        <activity android:name=".ForgotPasswordActivity" />
-        <activity android:name=".OrderActivity"/>
+        <activity android:name=".ui.navigation.SingUpActivity" />
+        <activity android:name=".ui.navigation.ForgotPasswordActivity" />
+        <activity android:name=".ui.navigation.OrderActivity"/>
     </application>
 
 </manifest>
