--- app/src/main/java/com/sanke46/android/e_commerce/SalesActivity.java
@@ -1,37 +1,0 @@
-package com.sanke46.android.e_commerce;
-
-import android.content.Intent;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
-import android.view.View;
-
-public class SalesActivity extends AppCompatActivity {
-
-    private DrawerLayout drawer;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sales);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        drawer = (DrawerLayout) findViewById(R.id.drawerLayout);
-
-        setSupportActionBar(toolbar);
-        getSupportActionBar().setTitle("Sales");
-
-
-        toolbar.setNavigationIcon(getResources().getDrawable(R.drawable.ic_arrow_back_black_36px));
-        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(getBaseContext(), MainActivity.class);
-                startActivity(intent);
-                finish();
-            }
-        });
-    }
-
-
-}
