+++ app/src/main/java/com/sanke46/android/e_commerce/adapter/ListAdapter.java
@@ -1,4 +1,4 @@
-package com.sanke46.android.e_commerce;
+package com.sanke46.android.e_commerce.adapter;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
@@ -10,17 +10,21 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.sanke46.android.e_commerce.R;
+import com.sanke46.android.e_commerce.model.Item;
+import com.sanke46.android.e_commerce.ui.navigation.BasketActivity;
+
 import java.util.List;
 
 /**
  * Created by ilafedoseev on 07.02.17.
  */
 
-public class ListAdapterBasket extends ArrayAdapter<Item>{
+public class ListAdapter extends ArrayAdapter<Item>{
 
     BasketActivity basketActivity = new BasketActivity();
     private List<Item> itemList = basketActivity.getBasketItem();
-    public ListAdapterBasket(Context context, List<Item> arrayList) {
+    public ListAdapter(Context context, List<Item> arrayList) {
         super(context, 0 ,arrayList);
     }
 
@@ -30,24 +34,26 @@ public class ListAdapterBasket extends ArrayAdapter<Item>{
     public View getView(final int position, View convertView, ViewGroup parent) {
         final Item item = getItem(position);
         if(convertView == null){
-            convertView = LayoutInflater.from(getContext()).inflate(R.layout.list_basket,parent,false);
+            convertView = LayoutInflater.from(getContext()).inflate(R.layout.list,parent,false);
         }
 
         ImageView image = (ImageView) convertView.findViewById(R.id.image);
         TextView name = (TextView) convertView.findViewById(R.id.name);
+        TextView comment = (TextView) convertView.findViewById(R.id.comment);
         TextView price = (TextView) convertView.findViewById(R.id.price);
-        Button delete = (Button) convertView.findViewById(R.id.delete);
+        Button buttonOne = (Button) convertView.findViewById(R.id.buttonOne);
+        Button buttonTwo = (Button) convertView.findViewById(R.id.buttonTwo);
 
         image.setImageResource(item.getImageId());
         name.setText(item.getName());
+        comment.setText(item.getComment());
         price.setText(item.getPrice() + " $");
-        delete.setOnClickListener(new View.OnClickListener() {
+        buttonTwo.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                itemList.remove(position);
+                System.out.println(getItemId(position));
+                itemList.add(getItem(position));
                 basketActivity.setBasketItem(itemList);
-//                basketActivity.;
-
             }
         });
