+++ app/src/main/java/com/sanke46/android/e_commerce/adapter/ListAdapterBasket.java
@@ -1,6 +1,5 @@
-package com.sanke46.android.e_commerce;
+package com.sanke46.android.e_commerce.adapter;
 
-import android.content.Context;
 import android.support.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -10,18 +9,25 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.sanke46.android.e_commerce.R;
+import com.sanke46.android.e_commerce.model.Item;
+import com.sanke46.android.e_commerce.ui.navigation.BasketActivity;
+
 import java.util.List;
 
 /**
  * Created by ilafedoseev on 07.02.17.
  */
 
-public class ListAdapter extends ArrayAdapter<Item>{
+public class ListAdapterBasket extends ArrayAdapter<Item>{
 
-    BasketActivity basketActivity = new BasketActivity();
-    private List<Item> itemList = basketActivity.getBasketItem();
-    public ListAdapter(Context context, List<Item> arrayList) {
+    BasketActivity basketActivity;
+    private List<Item> itemList;
+
+    public ListAdapterBasket(BasketActivity context, List<Item> arrayList) {
         super(context, 0 ,arrayList);
+        this.basketActivity = context;
+        this.itemList = basketActivity.getBasketItem();
     }
 
 
@@ -30,31 +36,28 @@ public class ListAdapter extends ArrayAdapter<Item>{
     public View getView(final int position, View convertView, ViewGroup parent) {
         final Item item = getItem(position);
         if(convertView == null){
-            convertView = LayoutInflater.from(getContext()).inflate(R.layout.list,parent,false);
+            convertView = LayoutInflater.from(getContext()).inflate(R.layout.list_basket,parent,false);
         }
 
         ImageView image = (ImageView) convertView.findViewById(R.id.image);
         TextView name = (TextView) convertView.findViewById(R.id.name);
-        TextView comment = (TextView) convertView.findViewById(R.id.comment);
         TextView price = (TextView) convertView.findViewById(R.id.price);
-        Button buttonOne = (Button) convertView.findViewById(R.id.buttonOne);
-        Button buttonTwo = (Button) convertView.findViewById(R.id.buttonTwo);
+        Button delete = (Button) convertView.findViewById(R.id.delete);
 
         image.setImageResource(item.getImageId());
         name.setText(item.getName());
-        comment.setText(item.getComment());
         price.setText(item.getPrice() + " $");
-        buttonTwo.setOnClickListener(new View.OnClickListener() {
+        delete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                itemList.remove(position);
+                basketActivity.refreshUi();
 
-                System.out.println(getItemId(position));
-                itemList.add(getItem(position));
-                basketActivity.setBasketItem(itemList);
+                notifyDataSetChanged();
+                notifyDataSetInvalidated();
             }
         });
 
         return convertView;
     }
-
 }
