+++ app/src/main/java/com/sanke46/android/e_commerce/database/DataBaseHandler.java
@@ -1,10 +1,12 @@
-package com.sanke46.android.e_commerce;
+package com.sanke46.android.e_commerce.database;
 
 import android.content.ContentValues;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+
+import com.sanke46.android.e_commerce.model.Item;
 
 import java.util.ArrayList;
 import java.util.List;
