+++ app/src/main/java/com/sanke46/android/e_commerce/ui/navigation/ProfileActivity.java
@@ -1,4 +1,4 @@
-package com.sanke46.android.e_commerce;
+package com.sanke46.android.e_commerce.ui.navigation;
 
 import android.content.Intent;
 import android.support.v4.widget.DrawerLayout;
@@ -6,6 +6,9 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
+
+import com.sanke46.android.e_commerce.MainActivity;
+import com.sanke46.android.e_commerce.R;
 
 public class ProfileActivity extends AppCompatActivity {
