+++ app/src/main/java/com/sanke46/android/e_commerce/ui/navigation/SalesActivity.java
@@ -1,30 +1,29 @@
-package com.sanke46.android.e_commerce;
+package com.sanke46.android.e_commerce.ui.navigation;
 
 import android.content.Intent;
-import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
+import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
 
-public class AboutDelevery extends AppCompatActivity {
+import com.sanke46.android.e_commerce.MainActivity;
+import com.sanke46.android.e_commerce.R;
+
+public class SalesActivity extends AppCompatActivity {
+
+    private DrawerLayout drawer;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_about_delevery);
+        setContentView(R.layout.activity_sales);
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
+        drawer = (DrawerLayout) findViewById(R.id.drawerLayout);
 
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
-            }
-        });
+        setSupportActionBar(toolbar);
+        getSupportActionBar().setTitle("Sales");
+
 
         toolbar.setNavigationIcon(getResources().getDrawable(R.drawable.ic_arrow_back_black_36px));
         toolbar.setNavigationOnClickListener(new View.OnClickListener() {
@@ -35,7 +34,7 @@ public class AboutDelevery extends AppCompatActivity {
                 finish();
             }
         });
-
     }
+
 
 }
