+++ app/src/main/res/layout/activity_about_delevery.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.sanke46.android.e_commerce.AboutDelevery">
+    tools:context="com.sanke46.android.e_commerce.ui.navigation.AboutDelevery">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
