+++ app/src/main/res/layout/activity_forgot_password.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.sanke46.android.e_commerce.ForgotPasswordActivity">
+    tools:context="com.sanke46.android.e_commerce.ui.navigation.ForgotPasswordActivity">
 
 </android.support.constraint.ConstraintLayout>
