+++ app/src/main/res/layout/activity_info_items.xml
@@ -8,6 +8,6 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.sanke46.android.e_commerce.InfoItemsActivity">
+    tools:context="com.sanke46.android.e_commerce.ui.navigation.InfoItemsActivity">
 
 </RelativeLayout>
