+++ app/src/main/res/layout/activity_profile.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.sanke46.android.e_commerce.ProfileActivity">
+    tools:context="com.sanke46.android.e_commerce.ui.navigation.ProfileActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
