+++ app/src/main/res/layout/activity_sales.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.sanke46.android.e_commerce.SalesActivity">
+    tools:context="com.sanke46.android.e_commerce.ui.navigation.SalesActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
