+++ app/src/main/res/layout/activity_sing_up.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.sanke46.android.e_commerce.SingUpActivity">
+    tools:context="com.sanke46.android.e_commerce.ui.navigation.SingUpActivity">
 
 </android.support.constraint.ConstraintLayout>
