+++ app/src/main/res/layout/content_about_delevery.xml
@@ -10,7 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.sanke46.android.e_commerce.AboutDelevery"
+    tools:context="com.sanke46.android.e_commerce.ui.navigation.AboutDelevery"
     tools:showIn="@layout/activity_about_delevery">
 
 </RelativeLayout>
