+++ app/src/main/res/layout/content_basket.xml
@@ -10,7 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.sanke46.android.e_commerce.BasketActivity"
+    tools:context="com.sanke46.android.e_commerce.ui.navigation.BasketActivity"
     tools:showIn="@layout/activity_basket">
 
 </RelativeLayout>
